MOD ecs/components/RotationComponent.hpp change type of the getter of the rotation

+++ ecs/components/RotationComponent.hpp
@@ -31,7 +31,7 @@ namespace ECS {
              * @brief Get the Rotation object
              * 
              */
-            int getRotation() const;
+            float getRotation() const;
             /**
              * @brief Set the Rotation object
              *
